+++ packages/front-end/components/Experiment/LinkedChanges/AddLinkedChanges.tsx
@@ -174,8 +174,8 @@ export default function AddLinkedChanges({
   };
 
   const possibleSections = Object.keys(sections);
-
   const sectionsToRender = possibleSections.filter((s) => sections[s].render);
+  if (!sectionsToRender.length) return null;
 
   return (
     <div className="appbox px-4 py-3 my-4">