+++ packages/back-end/src/services/experiments.ts
@@ -31,18 +31,22 @@ import {
   ExperimentMetricInterface,
   getAllMetricIdsFromExperiment,
   getEqualWeights,
+  getMetricResultStatus,
   getMetricSnapshotSettings,
   isFactMetric,
   isFactMetricId,
   isMetricJoinable,
 } from "shared/experiments";
-import { orgHasPremiumFeature } from "enterprise";
+import { orgHasPremiumFeature } from "shared/enterprise";
 import { hoursBetween } from "shared/dates";
 import { v4 as uuidv4 } from "uuid";
 import { MetricPriorSettings } from "back-end/types/fact-table";
 import {
+  ExperimentAnalysisSummaryVariationStatus,
   BanditResult,
   ExperimentAnalysisSummary,
+  ExperimentAnalysisSummaryResultsStatus,
+  GoalMetricResult,
 } from "back-end/src/validators/experiments";
 import { updateExperiment } from "back-end/src/models/ExperimentModel";
 import { promiseAllChunks } from "back-end/src/util/promise";
@@ -90,6 +94,7 @@ import {
 } from "back-end/types/experiment";
 import { findDimensionById } from "back-end/src/models/DimensionModel";
 import {
+  APP_ORIGIN,
   DEFAULT_CONVERSION_WINDOW_HOURS,
   EXPERIMENT_REFRESH_FREQUENCY,
 } from "back-end/src/util/secrets";
@@ -147,7 +152,12 @@ import {
   runSnapshotAnalysis,
   writeSnapshotAnalyses,
 } from "./stats";
-import { getEnvironmentIdsFromOrg } from "./organizations";
+import {
+  getConfidenceLevelsForOrg,
+  getEnvironmentIdsFromOrg,
+  getMetricDefaultsForOrg,
+  getPValueThresholdForOrg,
+} from "./organizations";
 
 export const DEFAULT_METRIC_ANALYSIS_DAYS = 90;
 
@@ -1320,6 +1330,8 @@ export async function toExperimentApiInterface(
   context: ReqContext | ApiReqContext,
   experiment: ExperimentInterface
 ): Promise<ApiExperiment> {
+  const appOrigin = (APP_ORIGIN ?? "").replace(/\/$/, "");
+
   let project: ProjectInterface | null = null;
   const organization = context.org;
   if (experiment.project) {
@@ -1335,6 +1347,7 @@ export async function toExperimentApiInterface(
   return {
     id: experiment.id,
     name: experiment.name || "",
+    type: experiment.type || "standard",
     project: experiment.project || "",
     hypothesis: experiment.hypothesis || "",
     description: experiment.description || "",
@@ -1417,6 +1430,12 @@ export async function toExperimentApiInterface(
             releasedVariationId: experiment.releasedVariationId || "",
             excludeFromPayload: !!experiment.excludeFromPayload,
           },
+        }
+      : null),
+    shareLevel: experiment.shareLevel || "organization",
+    ...(experiment.shareLevel === "public" && experiment.uid
+      ? {
+          publicUrl: `${appOrigin}/public/e/${experiment.uid}`,
         }
       : null),
   };
@@ -2253,7 +2272,7 @@ export function toMetricApiInterface(
         metric.maxPercentChange ?? metricDefaults?.maxPercentageChange ?? 0.5,
       minSampleSize:
         metric.minSampleSize ?? metricDefaults?.minimumSampleSize ?? 150,
-      targetMDE: metric.targetMDE ?? metricDefaults?.targetMDE ?? 0.01,
+      targetMDE: metric.targetMDE ?? metricDefaults?.targetMDE ?? 0.1,
       riskThresholdDanger: metric.loseRisk ?? 0.0125,
       riskThresholdSuccess: metric.winRisk ?? 0.0025,
       windowSettings: metric.windowSettings
@@ -2429,6 +2448,7 @@ export function postExperimentApiPayloadToInterface(
     regressionAdjustmentEnabled:
       payload.regressionAdjustmentEnabled ??
       !!organization?.settings?.regressionAdjustmentEnabled,
+    shareLevel: payload.shareLevel,
   };
 }
 
@@ -2468,6 +2488,7 @@ export function updateExperimentApiPayloadToInterface(
     statsEngine,
     regressionAdjustmentEnabled,
     secondaryMetrics,
+    shareLevel,
   } = payload;
   return {
     ...(trackingKey ? { trackingKey } : {}),
@@ -2541,6 +2562,7 @@ export function updateExperimentApiPayloadToInterface(
           }),
         }
       : {}),
+    ...(shareLevel !== undefined ? { shareLevel } : {}),
     dateUpdated: new Date(),
   };
 }
@@ -2836,6 +2858,26 @@ export async function updateExperimentAnalysisSummary({
     }
   }
 
+  const relativeAnalysis = experimentSnapshot.analyses.find(
+    (v) => v.settings.differenceType === "relative"
+  );
+
+  if (relativeAnalysis) {
+    const overallResults = relativeAnalysis.results?.[0];
+    // redundant check for dimension
+    if (overallResults && overallResults.name === "") {
+      const resultsStatus = await computeResultsStatus({
+        context,
+        relativeAnalysis,
+        experiment,
+      });
+
+      if (resultsStatus) {
+        analysisSummary.resultsStatus = resultsStatus;
+      }
+    }
+  }
+
   await updateExperiment({
     context,
     experiment,
@@ -2843,4 +2885,95 @@ export async function updateExperimentAnalysisSummary({
       analysisSummary,
     },
   });
+}
+
+async function computeResultsStatus({
+  context,
+  relativeAnalysis,
+  experiment,
+}: {
+  context: ReqContext;
+  relativeAnalysis: ExperimentSnapshotAnalysis;
+  experiment: ExperimentInterface;
+}): Promise<ExperimentAnalysisSummaryResultsStatus | undefined> {
+  const statsEngine = relativeAnalysis.settings.statsEngine;
+  const { ciUpper, ciLower } = getConfidenceLevelsForOrg(context);
+  const metricDefaults = getMetricDefaultsForOrg(context);
+  const pValueThreshold = getPValueThresholdForOrg(context);
+  const metricMap = await getMetricMap(context);
+
+  const variations = relativeAnalysis.results[0]?.variations;
+  if (!variations || !variations.length) {
+    return;
+  }
+
+  const variationStatuses: ExperimentAnalysisSummaryVariationStatus[] = [];
+  const baselineVariation = variations[0];
+
+  for (let i = 1; i < variations.length; i++) {
+    // try to get id from experiment object
+    const variationId = experiment.variations?.[i]?.id;
+    const currentVariation = variations[i];
+    const variationStatus: ExperimentAnalysisSummaryVariationStatus = {
+      variationId: variationId ?? i + "",
+      goalMetrics: {},
+      guardrailMetrics: {},
+    };
+    for (const m in currentVariation.metrics) {
+      const goalMetric = experiment.goalMetrics.includes(m);
+      const guardrailMetric = experiment.guardrailMetrics.includes(m);
+      if (goalMetric || guardrailMetric) {
+        const baselineMetric = baselineVariation.metrics?.[m];
+        const currentMetric = currentVariation.metrics?.[m];
+        if (!currentMetric || !baselineMetric) continue;
+
+        const metric = metricMap.get(m);
+        if (!metric) continue;
+
+        const resultsStatus = getMetricResultStatus({
+          metric,
+          metricDefaults,
+          baseline: baselineMetric,
+          stats: currentMetric,
+          ciLower,
+          ciUpper,
+          pValueThreshold,
+          statsEngine: statsEngine,
+        });
+
+        if (goalMetric) {
+          const metricStatus: GoalMetricResult = {
+            status: "neutral",
+            superStatSigStatus: "neutral",
+          };
+          if (resultsStatus.resultsStatus === "won") {
+            metricStatus.status = "won";
+          } else if (resultsStatus.resultsStatus === "lost") {
+            metricStatus.status = "lost";
+          }
+
+          if (resultsStatus.clearSignalResultsStatus === "won") {
+            metricStatus.superStatSigStatus = "won";
+          } else if (resultsStatus.resultsStatus === "lost") {
+            metricStatus.superStatSigStatus = "lost";
+          }
+          variationStatus.goalMetrics[metric.id] = metricStatus;
+        }
+
+        if (guardrailMetric) {
+          variationStatus.guardrailMetrics[metric.id] = {
+            status: resultsStatus.resultsStatus === "lost" ? "lost" : "neutral",
+          };
+        }
+      }
+    }
+    variationStatuses.push(variationStatus);
+  }
+
+  return {
+    variations: variationStatuses,
+    settings: {
+      sequentialTesting: relativeAnalysis.settings.sequentialTesting ?? false,
+    },
+  };
 }