+++ packages/back-end/types/organization.d.ts
@@ -12,7 +12,7 @@ import {
   CommercialFeature,
   LicenseInterface,
   SubscriptionInfo,
-} from "enterprise";
+} from "shared/enterprise";
 import { environment } from "back-end/src/routers/environment/environment.validators";
 import type { ReqContextClass } from "back-end/src/services/context";
 import { attributeDataTypes } from "back-end/src/util/organization.util";
@@ -243,7 +243,7 @@ export interface OrganizationSettings {
   requireExperimentTemplates?: boolean;
   experimentMinLengthDays?: number;
   experimentMaxLengthDays?: number;
-  midExperimentPowerEnabled?: boolean;
+  decisionFrameworkEnabled?: boolean;
 }
 
 export interface OrganizationConnections {