+++ packages/front-end/components/Ideas/ImpactModal.tsx
@@ -92,6 +92,7 @@ const ImpactModal: FC<{
           experimentLength,
           estimateParams: {
             estimate: est.id,
+            segment: est?.segment || "",
             improvement: value.improvement,
             numVariations: value.numVariations,
             userAdjustment: value.userAdjustment,