+++ packages/sdk-js/CHANGELOG.md
@@ -15,7 +15,7 @@
 ## **1.3.1** - Dec 3, 2024
 
 - Renamed `GrowthBookMultiUser` to `GrowthBookClient`
-- New `GrowthBookClient.getScopedInstance()` method to make it easier to use in express-like back-end frameworks and client-side environments.
+- New `GrowthBookClient.createScopedInstance()` method to make it easier to use in express-like back-end frameworks and client-side environments.
 
 ## **1.3.0** - Nov 20, 2024