+++ packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## **1.4.1** - Feb 20, 2025
+
+- In `auto.min.js`, enable dev mode by default. Without this, the GrowthBook DevTools Chrome extension will only partially work.
+
 ## **1.4.0** - Feb 19, 2025
 
 - Fixed edge case with pre-requisites that caused some feature rules to be skipped
@@ -11,7 +15,7 @@
 ## **1.3.1** - Dec 3, 2024
 
 - Renamed `GrowthBookMultiUser` to `GrowthBookClient`
-- New `GrowthBookClient.getScopedInstance()` method to make it easier to use in express-like back-end frameworks and client-side environments.
+- New `GrowthBookClient.createScopedInstance()` method to make it easier to use in express-like back-end frameworks and client-side environments.
 
 ## **1.3.0** - Nov 20, 2024